backend/requirements.txt
azure-identity==1.12.0
Flask==2.2.2
langchain==0.0.123
openai==0.27.2
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
msrestazure
azure.identity
azure.storage
azure.core
requests
azure-cognitiveservices-speech==1.22.0cd 
python-dotenv
azure-ai-formrecognizer==3.2.1
pypdf==3.7.1